- new logging location for easy access - add new status badge for pylint and pytest in documentation
